requirements.txt
boto3
dask
s3fs
pandas
openai
psycopg2-binary
streamlit